Remove toLowerCase to prevent conflicts
Public tags are not lowercased when created. This allows users to create multiple tags with the same name but which have different case: https://gyazo.com/c09f04ce7e80919397af69e57fde6fdd
In order to prevent conflicts, the toLowerCase call is removed when caching them.

Thanks again to RagingLink#9469 for the feedback!